refactor: move pagination provider
