+++ apps/backend/app/domains/ai/api/admin_quests_jobs_cursor_router.py
@@ -8,7 +8,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 
 from app.api.deps import admin_required
-from app.core.pagination import (
+from app.domains.ai.infrastructure.models.generation_models import GenerationJob
+from app.providers.db.pagination import (
     FilterSpec,
     apply_filters,
     apply_pagination,
@@ -19,7 +20,6 @@ from app.core.pagination import (
     fetch_page,
     parse_page_query,
 )
-from app.domains.ai.infrastructure.models.generation_models import GenerationJob
 from app.providers.db.session import get_db
 
 router = APIRouter(prefix="/admin/ai/quests", tags=["admin-ai-quests"])
