+++ apps/backend/app/providers/db/__init__.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 __all__ = [
     "adapters",
     "base",
+    "pagination",
     "sa_adapters",
     "session",
     "transition_query",
